requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-dateutil==2.8.1
python-dotenv==0.17.1
python-http-client==3.3.2
requests==2.25.1
sendgrid==6.7.0
six==1.16.0
soupsieve==2.2.1
starkbank-ecdsa==1.1.0
urllib3==1.26.4
